@@ -1,5 +1,5 @@
 import { NgModule } from '@angular/core';
-import { NgChartsModule } from 'ng2-charts';
+import { BaseChartDirective } from 'ng2-charts';
 import { CommonModule } from '@angular/common';
 import { FormsModule } from '@angular/forms';
 
@@ -9,6 +9,9 @@ import { SharedModule } from 'src/app/shared/shared.module';
 import { EstatisticasPageRoutingModule } from './estatisticas-routing.module';
 
 import { EstatisticasPage } from './estatisticas.page';
+import { Chart, PieController, ArcElement, Tooltip, Legend } from 'chart.js';
+Chart.register(PieController, ArcElement, Tooltip, Legend);
+
 
 @NgModule({
   imports: [
@@ -17,7 +20,7 @@ import { EstatisticasPage } from './estatisticas.page';
     IonicModule,
     SharedModule,
     EstatisticasPageRoutingModule,
-    NgChartsModule
+    BaseChartDirective
   ],
   declarations: [EstatisticasPage]
 })
